backend/requirements.txt
numpy==1.24.1
scipy==1.10.1
pillow==10.1.0
torch==2.1.1
accelerate==0.20.3
transformers==4.33.3
diffusers==0.21.4
safetensors==0.3.1
huggingface-hub==0.16.4
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
websockets==12.0 